requirements.txt
discord.py==2.3.2
httpx==0.27.2
python-dotenv==1.0.1
